requirements.txt
asgiref==3.6.0
Django==4.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
python-dotenv==1.0.0
pytz==2023.3
pytest==7.3.1
pytest-django==4.5.2
sqlparse==0.4.4
flake8==6.0.0
flake8-django==1.2
pyflakes==3.0.1


